requirements.txt
flask
sentence-transformers
torch
python-dotenv